+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ not yet released
 * #54: Improved phone/tablet compatibility.
 * #62: Replace javascript confirm dialogs with Bootstrap modal dialogs.
 * #66: Unable to add users with MySQL in strict mode.
+* #75: Test mode for email and SMS settings.
+* Composer added for dependencies.
 
 
 v3.0.1 (released April 12, 2014)